Move required fields into entity constructors
Previously we could get into a bad state where an entity object could be
created but could not be persisted if you had not called setProperty()
on a property that was marked as NOT NULL in the db schema but did not
have a sensible default set.

By creating constructors for our entities we can ensure the key
properties can be set when the object is instanciated.

* Require pid and title for CoreEntity
* Require pid and title for Image
* Require pid for Version
* Require type and name for VersionType
* Require type, version, programmeItem, scheduledStart and mediaSet for
  RefAvailablity
* Require name for RefMediaSet
* Require pid, subjectId, subjectType, objectId, objectType and
  relationshipType for RefRelationship
* Require pid and name for RefRelationshipType

This is a BREAKING CHANGE