Made Network required in the MasterBrand model
There's no value to having a MasterBrand without its Network as it is
almost always the Network entity that we care about.

In the cases where a MasterBrand in the DB doesn't have a Network
attached to it (in the short time between a MasterBrand being created
and the Network denorm running) we should consider that MasterBrand as
incomplete and not return any MasterBrand.