Add Segment/SegmentEvent domain models and mappers
Add set of Unfetched* Entities. When linking Entities with others there
are three states: "We've got the linked entity", "We requested the linked
entity and it came back null so we know there isn't one" and "We didn't
request the linked entity". Within our domain entities, the first is
represented by an instance of that entity, the second is represented by
null, and the third is represented by the Unfetched entities.

When you attempt to request a linked entity that hasn't been requested
at DataNotFetchedException is thrown signifying that you can't attempt
to render this information yet.