Overridden queryBuilders now leverage the Embargoed Filter
Previously we always explicitly removed Versions/SegmentEvents connected
to any embargoed items. Instead we should only do this when the embargoed
filter is enabled. This allows us to get results when we don't always
want to remove embargeod items (i.e. in Faucet).

Add tests for Version/SegmentEvent queries when the embargoed filter is
enabled and disabled.