Add LoggingServiceFactory
This is an attempt at providing logging for services without having to
litter function calls throughout all the service methods.

This is a new ServiceFactory that should be passed a logger and a
stopwatch in addition. It will then return proxy objects that return the
the services requested. Note that these are not properly typehinted (as
it is not possible without lots of evil eval magic). As such beware of
using them as they break any typehinting you might do where you expect
Service classes.