Cache mapper requests based on Database ID
Previously we always created entites on every call to getDomainModel()
however in many cases we had already constructed that model before, e.g.
if we are rendering a list of N children under a given parent then the
parent's domain model (and all the models that hang off it)  were being
created N times.

Now we cache the domain models in an array after creating them and when
a model is requested the second time we return the cached model.

All Mappers now use this process using the database ID, with the
exception of the CollapsedBroadcast which we can't create a unique cache
key for as it is a rolled up representation of many objects.

This is implemented within each Mapper rather than in an AbstractMapper
/ Trait because keeping the per-mapper return typehinting of
getDomainModel is useful and we would lose that if we used an
inheritance or composition based approach.