Add PartnerPid to entities that can have that set
In PIPs a PartnerPid may be null (as the concept was introduced after
PIPs' creation) but this is treated the same as if the entity has the
Partner Pid of 's0000001' - The BBC Partner Pid. In order to avoid two
states that do the same thing we shall set PartnerPid to be NOT NULL and
use a default value of 's0000001'.