Rework caching created domain objects
Previously the caching was naive and assumed the DB primary key would be
enough to distinguish objects apart. This was wrong as we can request
the same object with differing levels of hydration - e.g. a Programme
joined to the Image table and not, which would result in a Programme
with an UnfetchedImage.

This means that it was possible to request an entity without
relationships, then request it again with the relationships and the
second request would have an Unfetched entity incorrectly attached to
it.

To fix this we need to know the heierarchy of relationships and all
foreign keys when creating the cache key for a domain object.