Add CollapsedBroadcast Entity
This rolls up a multiple Broadcasts into a single entity, based on the
programme and start date. e.g. if a Programme was broadcast on
bbc_one_london and bbc_one_yorkshire at the same time, those two
broadcasts become a single CollapsedBroadcast. This also runs across
network so if something was broadcast on bbc_radio_foyle and
bbc_radio_ulster at the same time then they shall also result in a
single CollapsedBroadcast.

This will roll up different versions broadcast starting at the same time
(e.g. if a standard length and a shortened version are broadcast at the
same time they in will result in a single Collapsed broadcast. In this
case the shortest end_date and duration shall be stored.

This is possible using GROUP_CONCAT queries but is considerably slower,
so instead we'll do that denormalisation up front in Faucet.