Add option to inherit MasterBrand from parent CoreEntity
It is possible that a CoreEntity that is not relatated to a MasterBrand
does have a parent that is related to a MasterBrand, for example p043m65r

In the new frontend we want these Entities that do not have an explict
MasterBrand set to try and find one from their ancestry, in the same way
that we try and inherit images.

Clifton however does not do this inheritance.

In order to not make changes to Clifton's output, add an option
"core_entity_inherit_master_brand" to the MapperFactory that allows us
to choose between these two behaviours. When consuming PagesService you
probably want to set this option to true. We have not set this to be the
default as that would constitiute a breaking change, that we don't want
to make right now.