Allow RelatedLinksService->findByRelatedToProgramme() filter by link type
Allow for passing in an array of related link types.

I don't think filter by type requires an index, as we're already
filtering by the related_to id, which reduces the result set down to
less than 50 results in all cases (and less than 20 results in all but
45 cases).

Fixes PROGRAMMES-6221.