Calculate isExternal within the RelatedLink domain object
The external field from PIPs doesn't seem to be used as diligently as we
hoped. Instead work out if a Related Link is external or not on our own.

A link is considered external if the url is on bbc.co.uk or bbc.com or
one of their subdomains.

This is done lazily so that we can try and avoid regex calls to parse
the url if they are not needed.

Add getHost to RelatedLink.